+++ src/main/java/de/sakurajin/evenbetterarcheology/api/block/FossilCraftableWithEntity.java
@@ -4,16 +4,9 @@ import net.minecraft.block.*;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.screen.NamedScreenHandlerFactory;
 import net.minecraft.state.property.DirectionProperty;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.Hand;
-import net.minecraft.util.hit.BlockHitResult;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.shape.VoxelShape;
-import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
@@ -37,20 +30,4 @@ public abstract class FossilCraftableWithEntity extends FossilCraftable implemen
         if(expectedType != givenType) return null;
         return (BlockEntityTicker<A>) ticker;
     }
-
-    //Creates the Screen-Handler belonging to the BlockEntity
-    @SuppressWarnings("deprecation")
-    @Override
-    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
-        if (!world.isClient) {
-            NamedScreenHandlerFactory handledScreen = state.createScreenHandlerFactory(world, pos);
-
-            if (handledScreen != null) {
-                player.openHandledScreen(handledScreen);
-            }
-        }
-
-        return ActionResult.SUCCESS;
-    }
-
 }