+++ src/main/java/de/sakurajin/evenbetterarcheology/block/fossils/blockEntity/VillagerFossilBlockEntity.java
@@ -1,26 +1,19 @@
 package de.sakurajin.evenbetterarcheology.block.fossils.blockEntity;
 
-import de.sakurajin.evenbetterarcheology.block.entity.ImplementedInventory;
+import de.sakurajin.evenbetterarcheology.api.block.entity.BlockEntityWithInventory;
 import de.sakurajin.evenbetterarcheology.registry.ModBlockEntities;
 import de.sakurajin.evenbetterarcheology.block.fossils.VillagerFossilFull;
-import de.sakurajin.evenbetterarcheology.registry.ModMessages;
+import de.sakurajin.evenbetterarcheology.registry.ModNetworking;
 import de.sakurajin.evenbetterarcheology.screen.FossilInventoryScreenHandler;
-import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
-import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.Inventories;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
-import net.minecraft.network.PacketByteBuf;
 import net.minecraft.screen.NamedScreenHandlerFactory;
 import net.minecraft.screen.ScreenHandler;
-import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.Text;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
@@ -28,10 +21,9 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.Objects;
 
-public class VillagerFossilBlockEntity extends BlockEntity implements NamedScreenHandlerFactory, ImplementedInventory {
-    private final DefaultedList<ItemStack> inventory = DefaultedList.ofSize(1, ItemStack.EMPTY);
+public class VillagerFossilBlockEntity extends BlockEntityWithInventory implements NamedScreenHandlerFactory  {
     public VillagerFossilBlockEntity(BlockPos pos, BlockState state) {
-        super(ModBlockEntities.VILLAGER_FOSSIL, pos, state);
+        super(ModBlockEntities.VILLAGER_FOSSIL, pos, state, DefaultedList.ofSize(1, ItemStack.EMPTY));
     }
 
     @Override
@@ -47,11 +39,6 @@ public class VillagerFossilBlockEntity extends BlockEntity implements NamedScree
     }
 
     @Override
-    public DefaultedList<ItemStack> getItems(){
-        return this.inventory;
-    }
-
-    @Override
     public Text getDisplayName(){
         return Text.translatable(getCachedState().getBlock().getTranslationKey());
     }
@@ -59,8 +46,8 @@ public class VillagerFossilBlockEntity extends BlockEntity implements NamedScree
     //update luminance of block based on the luminance of the item given when it would be in its placed state
     @Override
     public void onClose(PlayerEntity player) {
-        ImplementedInventory.super.onClose(player);
-        int luminance = Block.getBlockFromItem(this.getInventoryContents().getItem()).getDefaultState().getLuminance();
+        super.onClose(player);
+        int luminance = Block.getBlockFromItem(this.getInventoryContents().get(0).getItem()).getDefaultState().getLuminance();
         player.getWorld().setBlockState(this.getPos(), Objects.requireNonNull(world).getBlockState(this.getPos()).with(VillagerFossilFull.INVENTORY_LUMINANCE, luminance));
     }
 
@@ -70,32 +57,10 @@ public class VillagerFossilBlockEntity extends BlockEntity implements NamedScree
         return new FossilInventoryScreenHandler(syncId, inv, this);
     }
 
-    public ItemStack getInventoryContents() {
-        return this.getStack(0);
-    }
-
-    public void setInventory(DefaultedList<ItemStack> inventory) {
-        for (int i = 0; i < inventory.size(); i++) {
-            this.inventory.set(i, inventory.get(i));
-        }
-    }
-
     @Override
     public void markDirty() {
-        assert world != null;
-        if(!world.isClient()) {
-            PacketByteBuf data = PacketByteBufs.create();
-            data.writeInt(inventory.size());
-            for (ItemStack itemStack : inventory) {
-                data.writeItemStack(itemStack);
-            }
-            data.writeBlockPos(getPos());
-
-            for (ServerPlayerEntity player : PlayerLookup.tracking((ServerWorld) world, getPos())) {
-                ServerPlayNetworking.send(player, ModMessages.ITEM_SYNC, data);
-            }
-
-            int luminance = Block.getBlockFromItem(this.getInventoryContents().getItem()).getDefaultState().getLuminance();
+        if(world != null && !world.isClient()) {
+            int luminance = Block.getBlockFromItem(this.getInventoryContents().get(0).getItem()).getDefaultState().getLuminance();
             world.setBlockState(this.getPos(), getCachedState().with(VillagerFossilFull.INVENTORY_LUMINANCE, luminance));
         }