+++ src/main/java/de/sakurajin/evenbetterarcheology/screen/IdentifyingScreenHandler.java
@@ -21,12 +21,12 @@ public class IdentifyingScreenHandler extends ScreenHandler {
     public IdentifyingScreenHandler(int syncId, PlayerInventory inventory) {
         //size of SimpleInventory has to be same as in Defaulted List of ArcheologyTableBlockEntity;
         //size of ArrayPropertyDelegate has to be the same as the PropertyDelegate of ArcheologyTableBlockEntity (number of ints being tracked)
-        this(syncId, inventory, new SimpleInventory(ArcheologyTableBlockEntity.INV_SIZE), new ArrayPropertyDelegate(ArcheologyTableBlockEntity.PROPERTY_DELEGATES));
+        this(syncId, inventory, new SimpleInventory(3), new ArrayPropertyDelegate(ArcheologyTableBlockEntity.PROPERTY_DELEGATES));
     }
 
     public IdentifyingScreenHandler(int syncId, PlayerInventory playerInventory, Inventory inventory, PropertyDelegate delegate) {
         super(ModScreenHandlers.IDENTIFYING_SCREEN_HANDLER, syncId);    //creates a new Instance of Screenhandler
-        checkSize(inventory, ArcheologyTableBlockEntity.INV_SIZE);
+        checkSize(inventory, 3);
         this.inventory = inventory; //sets the Screens Inventory to the given Inventory
 
         //opens Inventory